fix: cast capacity to uint32_t in allocator block bin lookups
- Apply explicit uint32_t cast before ceil_log2 in c_nt_shm, c_heap, and c_shm bin helpers
- Normalize em dashes to hyphens in related comments

+++ cbase/allocator_protocol/c_heap_allocator.h
@@ -128,7 +128,7 @@ static inline size_t c_heap_block_bin(size_t capacity) {
     if (capacity <= (size_t) AP_HEAP_EXACT_BIN_COUNT * 8u) {
         return capacity >> 3;
     }
-    size_t idx = (size_t) AP_HEAP_EXACT_BIN_COUNT + (c_heap_block_ceil_log2(capacity) - 16u);
+    size_t idx = (size_t) AP_HEAP_EXACT_BIN_COUNT + (c_heap_block_ceil_log2((uint32_t) capacity) - 16u);
     if (idx > (size_t) AP_HEAP_EXACT_BIN_COUNT + AP_HEAP_LARGE_BIN_COUNT) {
         idx = (size_t) AP_HEAP_EXACT_BIN_COUNT + AP_HEAP_LARGE_BIN_COUNT;
     }
@@ -367,7 +367,7 @@ static inline void* c_heap_request(heap_allocator* allocator, size_t size, int s
     }
 
     /* Step 1: size-binned free-list reuse (see c_heap_block_bin).
-     * Exact bins always fit (capacity == cap_net) — O(1) head pop.
+     * Exact bins always fit (capacity == cap_net) - O(1) head pop.
      * Pow2-class bins use in-bin first-fit: bins are short and same-
      * shaped, so the head is the common hit; larger-capacity blocks in
      * the bin also satisfy smaller requests. */
